move EqualsVerifier to base-crypto testFixtures
