+++ platform-sdk/consensus-hashgraph-impl/build.gradle.kts
@@ -11,6 +11,7 @@ description = "Default Consensus Hashgraph Implementation"
 testModuleInfo {
     requires("org.hiero.consensus.gui")
     requires("com.swirlds.base.test.fixtures")
+    requires("org.hiero.base.crypto.test.fixtures")
     requires("org.hiero.base.utility.test.fixtures")
     requires("com.swirlds.config.extensions.test.fixtures")
     requires("com.swirlds.platform.core.test.fixtures")
