+++ platform-sdk/consensus-model/build.gradle.kts
@@ -10,6 +10,7 @@ description = "Consensus Model"
 
 testModuleInfo {
     requires("com.swirlds.platform.core.test.fixtures")
+    requires("org.hiero.base.crypto.test.fixtures")
     requires("org.hiero.base.utility.test.fixtures")
     requires("org.hiero.consensus.model.test.fixtures")
     requires("org.hiero.consensus.utility")
