+++ platform-sdk/consensus-model/src/test/java/org/hiero/consensus/model/event/PlatformEventTest.java
@@ -10,13 +10,13 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import com.hedera.hapi.node.base.Timestamp;
 import com.hedera.hapi.platform.event.EventConsensusData;
 import com.hedera.hapi.platform.event.GossipEvent;
-import com.swirlds.platform.test.fixtures.utils.EqualsVerifier;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.time.Instant;
 import java.util.Random;
+import org.hiero.base.crypto.test.fixtures.EqualsVerifier;
 import org.hiero.base.io.streams.SerializableDataInputStream;
 import org.hiero.base.io.streams.SerializableDataOutputStream;
 import org.hiero.base.utility.test.fixtures.RandomUtils;
